search: Add the changed callback
This callback is simply responsible for updating the usability of the
categories sidebar, and then resetting the filter on the apps listbox.
It is then up to the filter function to take care of the rest. In the
next commit we'll move the filter function into the menu-search.c
file.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>