category: Also store a reference to the tree itself
This allows filtering contents of category based on the tree itself.
Note that we don't assign any group to the root tree "all" to allow
us to show all applications.

At some point we may wish to add support for more tree directories to
the menu itself so this makes a lot of sense.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>