window: Break focus now by becoming a popup type window
Popup windows follow different input requirements, and do not permanently
steal focus from other windows. For our menu window, we want to temporarily
hijack the keyboard and pointer devices, while not stealing the focus chain
from other windows.

This means we'll need to use X11/Seat grabbing to take exclusive ownership
when the popup window is mapped, so that we get all events.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>