Fix the issue of linking by providing our own .ctor and .dtor
The resources are built into a separate .a archive to prevent spammy
compiler warnings, but we link with -Wl,--no-whole-archive, which means
the generated .ctor's, .dtor's and resource section are completely absent.

We simplify it by calling them from the main executable to satisfy linking,
and thus, resource presence, while managing them in a .ctor/.dtor fashion.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>