entry: Immediately hide the window once a drag op begins
There is a 2-part issue here that is resolved. The first is obviously that
the menu itself can be in the way when wanting to drag icons out to the
desktop, so we shift out of view immediately.

The second more subtle issue is that once a drag operation begins, we then
lose our grab on the window. This invalidates the modal ability of the
toplevel, breaking input and such. So as a semi-workaround, instead of
trying to figure out *when* to restore grab, we can just safely hide
as we don't *accept* drops.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>