entry: Unbreak drag & drop
Under Marco, the ungrab happens considerably faster than under other
window managers, leading to a break on the dnd grab. To account for
this we'll now only ungrab/hide once the drag has ended, regardless
of whether it succeeded or not.

We also do this in the idle loop to ensure that our hide/ungrab doesn't
conflict with the current ongoing D&D operation.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>