sort: Ensure we sort in a case insensitive fashion
The prior sort algorithm is effectively a sort on the ASCII value of a
given character, thus A B will sort before a b, meaning "Df" appears before
"da".

This resolves issue #14 by ensuring we lower case the comparative strings
to ensure there is a logical case insensitive sort function.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>