entry: Only use GDesktopAppInfo
We now mandate that a valid GDesktopAppInfo is found for every entry
that we're attempting to add to the menu. If it doesn't load, i.e. because
it is missing, broken, has missing binaries, etc, we skip it completely.

Additionally, we now only use the GAppInfo icon API to ensure we have
a proper GThemedIcon, and internally allowing glib to return absolute
path icons and pixmap tricks to make all icons work.

This resolves issue #9.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>