window: Reset focus between shows
Resolve #5 by ensuring that the search focus is always restored when
showing the window again, i.e. because it was deselected in the tab
focus cycle on the previous show.

Additionally, we ensure we also reset any selection on the apps box that
may have existed, so that the window is presented in a pristine state every
time.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>