settings: Add support for configurable search position with settings
We now couple the applet and menu closer by adding a new gsettings key
for the position of the search bar, allowing top, bottom, and automatic.
We continue to use the previous behaviour as the default, by using a
top search bar.

However, when using "automatic" mode, the search bar will appear closest
to the panel. So using a top panel means a top search, and a bottom panel
means a bottom searchbar.

This resolves issue #2. Note that a UI for settings is currently out of
scope for v0.3.0

Signed-off-by: Ikey Doherty <ikey@solus-project.com>