applet: Add a new context menu for editing the menu structure
This change implements #16 by adding a new "Edit Menus" entry to the right
click (context) menu of the main applet on the panel. By default, we'll
attempt to load menulibre, but if we cannot find it, we'll fallback to mozo.

In both instances we'll make special care to fallback to binary names if
there is an issue locating the .desktop file, matching the upstream behaviour
in MATE panel.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>