applet: Fix large leak in new hotkey handling
This commit switches the code to respect the g_settings_get_string returning
a non-const (allocated) string, which must in turn be cleaned up upon dispose
and while switching shortcuts, otherwise we're leaking memory into the ether.

It also somewhat simplifies the callback for settings changes to only care
about one key, and doing so in just one function.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>