loader: Skip empty categories to fix #39
This is an interim method to skip any empty categories and not add them to
the sidebar. While users won't even notice the difference in initial load
speed, I'm not entirely happy with the lack of efficiency in this particular
check.

In future, when Brisk moves to a "data source" model, we'll not directly
interact with libmate-menus from the UI, so we'll be able to automatically
filter out any garbage.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>