backend/apps: Don't hold references to the trees
We'll load the trees on demand, and then purge them again so that
the library uncaches them. Our main task is to walk the tree and build
items for them, and drop all mate-menus references.

We'll add an idle reload mechanism which will take care of monitoring
the changes through GAppMonitor, as the internals of mate-menus has a
number of issues when it comes to handling notifications of app changes.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>