backend/apps: Add a new BriskAppsItem class
The BriskAppsItem is a .desktop file based BriskItem, which displays all
information directly from the corresponding .desktop file. As such, it
is a very very simple wrapper and makes life quite simple for us.

In future we'll pad this out with search + launch support, but for now
it is sufficient in terms of API design.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>