backend/apps: Emit fully formed objects for signal listeners to dup
We change the proprerty for the AppsItem to dupe the DesktopAppInfo, i.e.
we take a ref to it. This new AppsItem is then emitted to all listeners,
who can add it to the display, or ignore it. We always unref it ourselves,
however the frontend would ideally ref it and keep it for display.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>