backend/apps: Implement the new BriskAppsSection
The AppsSection will allow organising the items by section, and directly
corresponds to the sidebar categories in the menu. To ensure that we don't
hold any undue references on the MateMenuTreeDirectory, and thus, cause
issues with reloads, we cache all the properties we want and then nuke
it from orbit.

Note: As the BriskAppsItem currently doesn't expose any "parent" notion,
we don't yet implement the "can_show_item" filter function.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>