frontend: Initial port to the new backend system
This change is intentionally broken, it's just designed to let things
build right now. With this change we rewire all of frontend/ to stop
depending on mate-menus, and interface with the backend library instead.

The next steps will be to actually make Brisk use the new backend, and
to uncripple the functionality by moving bits around.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>