Through the power of magical floating references, we don't leak.
The GObject based params (and admittedly, some bugs) was causing an
ugly ref fight through signal races. We now ensure Item/Section are
GInitiallyUnowned, and that the buttons take the floating reference
to them. Once destroyed, the buttons in turn unref the floating ref
and all is well again.

In short we can now dynamically rebuild the menu list.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>