backend/apps: Collapse extra settings into 2 simple categories
The "mate-settings.menu" includes a structure which in turn includes
more categories, and then children. We now deliberately skip the intermediate
child section and group under the two main categories, chiefly:

 - Preferences
 - Administration

These are the two menus required by #10, and will be cleaned up and
reorganised a bit once we have separators + by-section filtering.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>