backend: Provide a new brisk_item_get_uri API for Drag&Drop
The EntryButton will now probe the corresponding Item to see if it
provides a usable (newly allocated) URL. If it does, we then set the
drag data to match.

Currently we only support .desktop files for DnD - but if the needs
arise in future we may look at supporting Places and such.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>