Restore the ability to launch things by clicking. Yknow, like a menu.
To ensure there is only one global context manager within Brisk, we
still make sure that the items are launched using our GdkAppLaunchContext.
This context is passed to the BriskItem, which it can then use to
launch (i.e. g_app_info_launch).

This maintains a clean sensible API with the bulk of the launch stuff
still happening in one place.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>