mate-applet: Split the existing positioning code out of the main applet
This is the first step towards supporting issue #50, by making the menu
placement code a bit more modular and split out from the main bulk of the
applet code, so we have somwhere clean & logical to work on it.

This change also makes the BriskMenuApplet type transparent so that we can
access the internals across multiple files. We have no reason to implement
an opaque applet type - we're not providing an API/ABI to external consumers
with Brisk anyway.

Additionally, we fold in the layout code as it primarily deals with orient
issues, which are close in context to the positioning code.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>