applet: Implement full support for vertical positioning
This resolves #50 by implementing a more intelligent window positioning
logic, that accounts for the positiong of the panel. When we find ourselves
in a vertical position, the menu will appear on the opposing edge and extend
outwards.

In horizontal layouts, we'll extend away from the the horizontal face of
the applet towards the center.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>