backend: Add a new special "All Items" backend
As part of issue #49 we add a new "All Items" backend, whose only job is
to emit a single section for the sidebar. This section in turn returns
TRUE for every single item possible, which magically results in an "All"
filter (by not filtering :P)

This is not currently hooked up to the backend, that will happen in the
next changes.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>