Fix Super key shortcuts being swallowed
This fixes an issue where shortcuts that use the Super key are being
swallowed by brisk-menu. This branch does not fix _all_ instances of the
issue, but a large subset of them seem to be fixed by it.

A workaround is to set any desired shortcuts that use the Super key
directly in the Compiz configuration (assuming you use Compiz).

This partially addresses the issues in #55.