backend/favourites: Make the backend use O(1) lookups
This changes the pin behaviour to use O(1) lookups and not O(N) to make
the backend a little bit more speedy. However this does expose the issue
that the Section is duplicating some behaviour of the Backend, so we'll
fix this in a follow up commit by exposing the owning backend to the section,
which will allow for private API to reduce duplication.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>