frontend: Install a new `show-context-menu` signal on menu buttons
In order to get the menu functionality back to being lightweight, we need
to have a single global menu handler for our items. The first step is to
add a new signal that the BriskMenuEntryButton will emit when it needs
a "right click menu" to be shown on screen.

The next steps will be to move this all to a global handler within frontend,
and then we'll look to remove the leaks and unnecessary extra allocations.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>