applet: Report errors when `Edit Menus` function fails
To make this more useful and less intrusive, we'll pop up a passive
notification to indicate that the user needs to install a menu editor
to use the function, instead of just doing nothing.

We must use libnotify to achieve this currently, because GNOME/GTK's
GNotification is bound entirely to GApplication, which we do not use
in Brisk (And indeed, cannot.)

This fixes #62.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>