Rewrite get_item_actions with action-group based menu structures
This ensures we have full accountancy and ownership throughout for the
menu structure and actions, ensuring it actually gets freed in the end.
Additionally this will actually allow each backend to submit their own
GMenuModel, and opens up the door for stateful actions, whilst keeping
the API clean and efficient.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>