Convert build system to meson
This is just going to make my life that much easier by not having to
fart about with autotools any more. Additionally, it helped to make
parts of the system more modular, such as splitting the libraries up
so that other components could link to them (i.e. new frontends!).

Signed-off-by: Ikey Doherty <ikey@solus-project.com>